Approaching (Almost) Any Machine Learning Problem
This book is for people who have some theoretical knowledge of machine learning and deep learning and want to dive into applied machine learning. The book is more oriented towards how and what should you use to solve machine learning and deep learning problems. The book is for you if you are looking for guidance on approaching machine learning problems